+++ ComercioDigital/ComercioDigital/DTOs/Productos/Moda/Joyeria.cs
@@ -11,7 +11,7 @@ namespace ComercioDigital.DTOs.Productos.Moda
     {
         public string Medida { get; set; }
 
-        public Joyeria(string nombre, string marca, float precio, Vendedor vendedor, string descripcion,
+        public Joyeria(string nombre, string marca, decimal precio, Vendedor vendedor, string descripcion,
             DateTime fechaPuestaVenta, string codigoDescuento, int stock, string color, string mAterial, string sexo,
             string medida) : base(nombre, marca, precio, vendedor, descripcion, fechaPuestaVenta, codigoDescuento,
             stock, color, mAterial, sexo)
