+++ ComercioDigital/ComercioDigital/DTOs/Productos/Moda/Ropa.cs
@@ -12,7 +12,7 @@ namespace ComercioDigital.DTOs.Productos.Moda
         public string Talla { get; set; }
         public string Tipo { get; set; }
 
-        public Ropa(string nombre, string marca, float precio, Vendedor vendedor, string descripcion,
+        public Ropa(string nombre, string marca, decimal precio, Vendedor vendedor, string descripcion,
             DateTime fechaPuestaVenta, string codigoDescuento, int stock, string color, string mAterial, string sexo,
             string talla, string tipo) : base(nombre, marca, precio, vendedor, descripcion, fechaPuestaVenta,
             codigoDescuento, stock, color, mAterial, sexo)
