+++ ComercioDigital/ComercioDigital/DTOs/Productos/Tecnologia/Ordenador.cs
@@ -13,7 +13,7 @@ namespace ComercioDigital.DTOs.Productos.Tecnologia
     public string PlacaBase { get; set; }
     public string Tipo { get; set; }
 
-    public Ordenador(string nombre, string marca, float precio, Vendedor vendedor, string descripcion,
+    public Ordenador(string nombre, string marca, decimal precio, Vendedor vendedor, string descripcion,
         DateTime fechaPuestaVenta, string codigoDescuento, int stock, string color, string procesador, string so,
         string modelo, DateTime fechaLanzamiento, string placaBase, string tipo) : base(nombre, marca, precio, vendedor,
         descripcion, fechaPuestaVenta, codigoDescuento, stock, color, procesador, so, modelo, fechaLanzamiento)
